More schema!
Spread the schema love around to a few other places.

There’s still room for improvement.

Ideally, at some point, posts should come with an option to choose
something more specific than an “Article” type (such as “blogPosting.”
As it stands, I used the more generic type to cover as many sites as
possible.

Full schema flexibility is a premium theme feature. Would be nice to
have it in a freebie!